requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.75.0
grpcio==1.45.0
grpcio-tools==1.45.0
h11==0.13.0
idna==3.3
motor==2.5.1
protobuf==3.19.4
pydantic==1.9.0
pymongo==3.12.3
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
yapf==0.32.0
